Moved socket connection from App Component to orders page component since it is the only one to require that specific connection
