+++ src/components/App.jsx
@@ -2,37 +2,19 @@ import React from 'react';
 import { useState, useEffect } from 'react';
 import io from 'socket.io-client';
 import axios from 'axios';
+import Board from './Board.jsx';
 
 // components
 
 const App = ({props}) => {
-  const [isConnected, setIsConnected] = useState(false);
-
-  useEffect(() => {
-    const socket = io('http://localhost:9000'); // url of the server is used to create socket
-    // when connected
-    socket.on('connect', (res) => {
-      setIsConnected('true');
-    });
-
-    socket.on('time', (time) => {
-      setIsConnected(time);
-    });
-
-    // use await to fetch dots
-
-    socket.on('disconnect', () => {
-      setIsConnected('Disconnected from Server');
-    });
-  }, []);
 
   return (
     <div>
       <h1> Restaurant Orders</h1>
-      It is {isConnected} that we are connected.
+      HTML SHOWING that we are connected.
+      <Board />
     </div>
   );
 }
 
 export default App;
-
